PR fixes and linting
