+++ AutoDraft/drift.py
@@ -1,0 +1,6 @@
+import pandas as pd
+
+@st.cache
+def load_csv(path='./data/full_dataset_4_seasons.csv'):
+    data = pd.read_csv(path)
+    return data
