--- data/assemble_multiplayer_stat_dataframe_TEMP.csv
@@ -1,2 +1,0 @@
-,name,gameNumber,season,statTimeonice,statAssists,statGoals,statPim,statShots,statGames,statHits,statPowerplaygoals,statPowerplaypoints,statPowerplaytimeonice,statEventimeonice,statPenaltyminutes,statShotpct,statGamewinninggoals,statOvertimegoals,statShorthandedgoals,statShorthandedpoints,statShorthandedtimeonice,statBlocked,statPlusminus,statPoints,statShifts,teamId,teamName,teamLink,opponentId,opponentName,opponentLink,date,isHome,isWin,isOT,gameLink,cumStatpoints
-0,Jimmy Schuldt,1,20182019,21:03,1,0,0,1,1,0,0,0,01:49,19:12,0,0.0,0,0,0,0,00:02,1,-1,1,22,54,Vegas Golden Knights,/api/v1/teams/54,26,Los Angeles Kings,/api/v1/teams/26,2019-04-06,False,False,False,/api/v1/game/2018021270/feed/live,1
